requirements.txt
# Voice Journey - Core Dependencies
# Install with: pip install -r requirements.txt
# 
# Note: For full environment reproduction, use conda_packages.txt instead:
#   conda install --file conda_packages.txt
#
# Date: December 21, 2025

# PyTorch (install separately for CUDA support)
# See: https://pytorch.org/get-started/locally/
# torch>=2.0.0
# torchaudio>=2.0.0

# Speaker diarization
pyannote-audio>=4.0.0

# Speech transcription
faster-whisper>=1.0.0

# Hugging Face integration
huggingface-hub>=0.20.0
transformers>=4.30.0

# Audio processing
soundfile>=0.12.0

# Utilities
tqdm>=4.60.0
